refactor: move auth options to separate module
