+++ app-main/app/protected/layout.tsx
@@ -1,5 +1,5 @@
 import { getServerSession } from "next-auth";
-import { authOptions } from "../api/auth/[...nextauth]/route";
+import { authOptions } from "../../lib/auth";
 import { redirect } from "next/navigation";
 
 export default async function ProtectedLayout({
