+++ app-main/lib/auth.ts
@@ -1,0 +1,42 @@
+import { NextAuthOptions, AuthOptions, Session } from "next-auth";
+import AzureADProvider from "next-auth/providers/azure-ad";
+import { JWT } from "next-auth/jwt";
+
+export const authOptions: NextAuthOptions = {
+  providers: [
+    AzureADProvider({
+      clientId: process.env.NEXT_PUBLIC_AZURE_AD_CLIENT_ID!,
+      clientSecret: process.env.AZURE_AD_CLIENT_SECRET!,
+      tenantId: process.env.NEXT_PUBLIC_AZURE_AD_TENANT_ID!,
+      checks: ["pkce", "state"],
+      authorization: {
+        params: {
+          scope:
+            "openid profile email offline_access api://ccd5ce0d-5f7a-43cf-bbb8-82c9cb822f12/access_as_user",
+        },
+      },
+    }),
+  ],
+  session: {
+    strategy: "jwt",
+  },
+  secret: process.env.NEXTAUTH_SECRET,
+  callbacks: {
+    async jwt({ token, account }: { token: JWT; account?: any }) {
+      if (account?.access_token) {
+        token.accessToken = account.access_token;
+      }
+      return token;
+    },
+    async session({ session, token }: { session: Session; token: JWT }) {
+      return {
+        ...session,
+        accessToken: token.accessToken,
+      };
+    },
+    async redirect({ url, baseUrl }) {
+      return `${baseUrl}/welcome`;
+    },
+  },
+  debug: true,
+};
